trekk ut flere utils og tag-komponent
