+++ apps/innbyggers-flate/src/pages/TiltakPage.tsx
@@ -3,31 +3,28 @@ import {
   BodyLong,
   BodyShort,
   ExpansionCard,
-  Heading,
   HStack,
+  Heading,
   Label,
   Link,
   LinkPanel,
   List
 } from '@navikt/ds-react'
 import {
+  DeltakerStatusTag,
   DeltakerStatusType,
+  EMDASH,
+  INNHOLD_TYPE_ANNET,
+  Tiltakstype,
+  formatDateFromString,
   getDeltakerStatusAarsakText,
-  Tiltakstype
+  hentTiltakNavnHosArrangorTekst
 } from 'deltaker-flate-common'
+import { useDeltakerContext } from '../DeltakerContext.tsx'
 import { DeltakerResponse } from '../api/data/deltaker.ts'
 import { DeltakerStatusInfoTekst } from '../components/DeltakerStatusInfoTekst.tsx'
-import { DeltakerStatusTag } from '../components/DeltakerStatusTag.tsx'
 import { HvaErDette } from '../components/HvaErDette.tsx'
-import { useDeltakerContext } from '../DeltakerContext.tsx'
 import { getDialogUrl } from '../utils/environment-utils.ts'
-import {
-  EMDASH,
-  formatDateFromString,
-  hentTiltakNavnHosArrangørTekst,
-  INNHOLD_TYPE_ANNET
-} from '../utils/utils.ts'
-
 const skalViseDeltakelsesmengde = (deltaker: DeltakerResponse) => {
   return (
     deltaker.deltakerliste.tiltakstype == Tiltakstype.ARBFORB ||
@@ -59,7 +56,7 @@ const skalViseDeltakerStatusInfoTekst = (status: DeltakerStatusType) => {
 export const TiltakPage = () => {
   const { deltaker } = useDeltakerContext()
 
-  const tiltakOgStedTekst = hentTiltakNavnHosArrangørTekst(
+  const tiltakOgStedTekst = hentTiltakNavnHosArrangorTekst(
     deltaker.deltakerliste.tiltakstype,
     deltaker.deltakerliste.arrangorNavn
   )
