+++ apps/nav-veileders-flate/src/components/rediger-pamelding/Utkast.tsx
@@ -1,7 +1,6 @@
 import { BodyLong, Heading, List, VStack } from '@navikt/ds-react'
-import { Tiltakstype } from 'deltaker-flate-common'
+import { EMDASH, INNHOLD_TYPE_ANNET, Tiltakstype } from 'deltaker-flate-common'
 import { Deltakelsesinnhold } from '../../api/data/pamelding.ts'
-import { EMDASH, INNHOLD_TYPE_ANNET } from '../../utils/utils.ts'
 
 interface Props {
   innhold: Deltakelsesinnhold | null
