+++ apps/nav-veileders-flate/src/utils/pamelding-form-utils.ts
@@ -5,7 +5,8 @@ import {
   SendInnPameldingRequest
 } from '../api/data/send-inn-pamelding-request.ts'
 import { SendInnPameldingUtenGodkjenningRequest } from '../api/data/send-inn-pamelding-uten-godkjenning-request.ts'
-import { DeltakelsesprosentValg, INNHOLD_TYPE_ANNET } from './utils.ts'
+import { DeltakelsesprosentValg } from './utils.ts'
+import { INNHOLD_TYPE_ANNET } from 'deltaker-flate-common'
 
 export const generateInnholdFromResponse = (
   pamelding: PameldingResponse,
@@ -45,7 +46,9 @@ export const generatePameldingRequestFromForm = (
     deltakerlisteId: pamelding.deltakerliste.deltakerlisteId,
     dagerPerUke: data.dagerPerUke,
     deltakelsesprosent:
-      data.deltakelsesprosentValg === DeltakelsesprosentValg.JA ? 100 : data.deltakelsesprosent,
+      data.deltakelsesprosentValg === DeltakelsesprosentValg.JA
+        ? 100
+        : data.deltakelsesprosent,
     bakgrunnsinformasjon: data.bakgrunnsinformasjon,
     innhold: generateInnholdFromResponse(
       pamelding,
