+++ packages/deltaker-flate-common/index.ts
@@ -1,5 +1,7 @@
+export * from './components/DeltakerStatusTag'
 export * from './hooks/useDeferredFetch'
 export * from './hooks/useFetch'
 export * from './model/deltaker'
 export * from './utils/displayText'
-// export * from './utils/utils'
+export * from './utils/constants'
+export * from './utils/utils'
