+++ packages/deltaker-flate-common/tailwind.config.ts
@@ -1,0 +1,2 @@
+import config from '@amt-deltaker-flate/tailwind-config/tailwind.config'
+export default config
