+++ packages/deltaker-flate-common/utils/constants.ts
@@ -1,0 +1,2 @@
+export const EMDASH = '—'
+export const INNHOLD_TYPE_ANNET = 'annet'
